requirements.txt
discord.py>=2.0.0
apscheduler>=3.9.1
pytz
python-dotenv
